fix task cores count
